Formated long numbers into more friendly string

+++ app/src/main/java/com/codigodelsur/androidsampleproject/model/Image.java
@@ -21,12 +21,12 @@ public class Image {
 
 
     @SerializedName("views")
-    public double favoritesCount;
+    public long favoritesCount;
 
     @SerializedName("comments")
-    public double commentsCount;
+    public long commentsCount;
 
     @SerializedName("downloads")
-    public double downloadsCount;
+    public long downloadsCount;
 
 }
